core: move EigenFromPy to a dedicated file
